feat(logging): add progress formatter
