+++ crates/cli/Cargo.toml
@@ -14,6 +14,7 @@ compress = { path = "../compress" }
 shell-words = "1.1"
 meta = { path = "../meta", default-features = false }
 daemon = { path = "../daemon" }
+logging = { path = "../logging" }
 
 [dev-dependencies]
 tempfile = "3"
