+++ crates/engine/Cargo.toml
@@ -13,6 +13,7 @@ nix = { version = "0.27", features = ["user", "fs"] }
 meta = { path = "../meta", default-features = false }
 blake3 = "1"
 fastcdc = "3.2"
+logging = { path = "../logging" }
 
 [dev-dependencies]
 tempfile = "3"
